backend/requirements.txt
duckduckgo-search>=6.0.0
playwright>=1.42.0
python-whois>=0.8.0
dnspython>=2.6.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
